% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asleep.R, R/py_asleep.R
\name{asleep}
\alias{asleep}
\alias{py_asleep}
\title{Run \code{asleep} Model on Data}
\usage{
asleep(
  file,
  outdir = NULL,
  min_wear_hours = 22L,
  time_shift = "0",
  report_light_and_temp = FALSE,
  pytorch_device = c("cpu", "cuda:0"),
  sample_rate = NULL,
  verbose = TRUE,
  force_download = FALSE
)

py_asleep(
  ...,
  pyenv_function = function() {
    
    reticulate::py_require(c("asleep @ git+https://github.com/muschellij2/asleep.git",
    "argparse", "numpy", "pandas", "importlib"), python_version = "3.8")
    
    reticulate::import("asleep")
 },
  show = FALSE
)
}
\arguments{
\item{file}{accelerometry file to process, including CSV,
CWA, GT3X, and \code{GENEActiv} bin files}

\item{outdir}{output directory for CSVs and outputs}

\item{min_wear_hours}{Min wear time in hours to be eligible for summary statistics computation.
The sleepnet paper uses 22}

\item{time_shift}{The number hours to shift forward or backward from
the current device time. e.g. +1 or -1}

\item{report_light_and_temp}{If true, it adds mean temp, and light columns to the predictions}

\item{pytorch_device}{device to use for prediction for PyTorch.}

\item{sample_rate}{Sample rate for the data, currently not used.}

\item{verbose}{print diagnostic messages}

\item{force_download}{force a download of the model, passed to \code{\link[=sl_download_models]{sl_download_models()}}}

\item{...}{arguments to pass to \link{asleep}}

\item{pyenv_function}{function that loads the forest Python package.
By default, it uses \code{reticulate::py_import("asleep")} to import
the package. If this function has an args argument, the output of
\code{pyenv_function} will be re-assigned to args.}

\item{show}{Logical, whether to show the standard output on the
screen while the child process is running, passed to \code{\link[callr:r]{callr::r()}}}
}
\value{
A list of outputs, including summaries, paths, and dataframes.
}
\description{
Run \code{asleep} Model on Data

Run \code{asleep} with Python
}
\examples{
asleep_check_result = function() {
  res = try({suppressWarnings(asleep_check())})
  if (inherits(res, "try-error")) {
    res = FALSE
  }
  res
}
\donttest{
  file = system.file("extdata/example_sleep.csv.gz", package = "asleep")
  stopifnot(file.exists(file))
  if (asleep_check_result()) {
    try({sl_download_models()})
    out = try({asleep(file = file, verbose = 2L)})
    if (inherits(out, "try-error")) {
      message(out)
      if (interactive()) {
        reticulate::py_last_error()
      }
    } else {
      pred = out$predictions
    }
  }
}
\donttest{
  file = system.file("extdata/example_sleep.csv.gz", package = "asleep")
  df = readr::read_csv(file)
  if (asleep_check_result()) {
    out = try({asleep(file = df)})
    if (inherits(out, "try-error")) {
      message(out)
    } else {
      st = out$predictions
      if (requireNamespace("ggplot2", quietly = TRUE) &&
          requireNamespace("tidyr", quietly = TRUE) &&
          requireNamespace("dplyr", quietly = TRUE)) {
        d = st[1:250,] \%>\%
          dplyr::mutate(
            time = lubridate::as_datetime(time),
            time_end = dplyr::lead(time)
          ) \%>\%
          dplyr::filter(!is.na(time_end))
        raw = df \%>\% dplyr::filter(time >= min(d$time) & time <= max(d$time))
        dat = raw \%>\%
          tidyr::gather(axis, value, -time)
        d = d \%>\% dplyr::mutate(activity_y = as.numeric(sleep_wake == "sleep"))
        dat \%>\%
          ggplot2::ggplot(ggplot2::aes(x = time, y = value, colour = axis)) +
          ggplot2::geom_step() +
          ggplot2::geom_segment(
            data = d,
            ggplot2::aes(
              x = time, xend = time_end,
              y = activity_y, yend = activity_y,
              linetype = sleep_wake
            ),
            colour = "black", linewidth = 1, inherit.aes = FALSE
          ) +
          ggplot2::labs(linetype = "Sleep/wake")
      }
    }
  }
}
}
