% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asleep.R
\name{have_asleep}
\alias{have_asleep}
\alias{asleep_check}
\alias{asleep_version}
\title{Check the \code{asleep} Python Module}
\usage{
have_asleep()

asleep_check(...)

asleep_version(...)
}
\arguments{
\item{...}{additional arguments to pass to \link[reticulate:py_list_packages]{reticulate::py_list_packages}}
}
\value{
A logical value indicating whether the \code{asleep} Python module is available.
}
\description{
Check the \code{asleep} Python Module
}
\examples{
\donttest{
  if (have_asleep()) {
     asleep_version()
  }
}
}
