% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_load_model.R
\name{sl_download_models}
\alias{sl_download_models}
\title{Download \code{asleep} models to Cache}
\usage{
sl_download_models(force_download = FALSE)
}
\arguments{
\item{force_download}{force a download of the model, even if it already exists at \code{model_path}}
}
\value{
A model from Python.  \code{sl_download_model} returns a model file path.
}
\description{
Download \code{asleep} models to Cache
}
