% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_load_model.R
\name{sl_load_model}
\alias{sl_load_model}
\alias{sl_download_model}
\title{Load \code{asleep} Model}
\usage{
sl_load_model(model_path = NULL, force_download = FALSE, as_python = TRUE)

sl_download_model(model_path, ...)
}
\arguments{
\item{model_path}{the file path to the model.  If on disk, this can be
re-used and not re-downloaded.  If \code{NULL}, will download to the
temporary directory}

\item{as_python}{Keep model object as a python object}

\item{...}{for \code{sl_download_model}, additional arguments to pass to
\code{\link[curl:curl_download]{curl::curl_download()}}}
}
\value{
A model from Python.  \code{sl_download_model} returns a model file path.
}
\description{
Load \code{asleep} Model
}
