% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_read.R
\name{sl_read}
\alias{sl_read}
\title{Read a Data Set for \code{asleep}}
\usage{
sl_read(file, resample_hz = "uniform", keep_pandas = FALSE)
}
\arguments{
\item{file}{path to the file for reading}

\item{resample_hz}{Target frequency (Hz) to resample the signal. If
"uniform", use the implied frequency (use this option to fix any device
sampling errors). Pass \code{NULL} to disable. Defaults to "uniform".}

\item{keep_pandas}{do not convert the data to a \code{data.frame} and keep
as a \code{pandas} \code{data.frame}}
}
\value{
A list of the data and information about the data
}
\description{
Read a Data Set for \code{asleep}
}
\note{
The data \code{P30_wrist100} is from
\url{https://ora.ox.ac.uk/objects/uuid:19d3cb34-e2b3-4177-91b6-1bad0e0163e7},
where we took the first 180,000 rows, the first 30 minutes of data
from that participant as an example.
}
\examples{

file = system.file("extdata/P30_wrist100.csv.gz", package = "asleep")
if (asleep_check()) {
  out = sl_read(file)
}
\dontrun{
  file = system.file("extdata/P30_wrist100.csv.gz", package = "asleep")
  if (asleep_check()) {
    out = sl_read(file, sample_rate = 100L)
  }
}
}
