% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{sl_rename_data}
\alias{sl_rename_data}
\alias{sl_write_csv}
\title{Rename data for \code{asleep}}
\usage{
sl_rename_data(data)

sl_write_csv(data, path = tempfile(fileext = ".csv"))
}
\arguments{
\item{data}{a \code{data.frame} of raw accelerometry}

\item{path}{path to the CSV output file}
}
\value{
A \code{data.frame} of renamed columns
}
\description{
Rename data for \code{asleep}
}
