% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asleep.R
\name{asleep}
\alias{asleep}
\title{Run \code{asleep} Model on Data}
\usage{
asleep(
  file,
  outdir = NULL,
  min_wear_hours = 22L,
  time_shift = "0",
  report_light_and_temp = FALSE,
  pytorch_device = c("cpu", "cuda:0"),
  sample_rate = NULL,
  verbose = TRUE,
  force_download = FALSE
)
}
\arguments{
\item{file}{accelerometry file to process, including CSV,
CWA, GT3X, and \code{GENEActiv} bin files}

\item{outdir}{output directory for CSVs and outputs}

\item{min_wear_hours}{Min wear time in hours to be eligible for summary statistics computation.
The sleepnet paper uses 22}

\item{time_shift}{The number hours to shift forward or backward from
the current device time. e.g. +1 or -1}

\item{report_light_and_temp}{If true, it adds mean temp, and light columns to the predictions}

\item{pytorch_device}{device to use for prediction for PyTorch.}

\item{sample_rate}{Sample rate for the data, currently not used.}

\item{verbose}{print diagnostic messages}

\item{force_download}{force a download of the model, passed to \code{\link[=sl_download_models]{sl_download_models()}}}
}
\value{
A list of outputs, including summaries, paths, and dataframes.
}
\description{
Run \code{asleep} Model on Data
}
\examples{
\donttest{
  file = system.file("extdata/example_sleep.csv.gz", package = "asleep")
  stopifnot(file.exists(file))
  if (asleep_check()) {
    sl_download_models()
    out = try({asleep(file = file, verbose = 2L)})
    if (inherits(out, "try-error")) {
      message(out)
      reticulate::py_last_error()
    } else {
      pred = out$predictions
    }
  }
}
\donttest{
  file = system.file("extdata/example_sleep.csv.gz", package = "asleep")
  df = readr::read_csv(file)
  if (asleep_check()) {
    out = asleep(file = df)
    st = out$predictions
  }
  if (requireNamespace("ggplot2", quietly = TRUE) &&
      requireNamespace("tidyr", quietly = TRUE) &&
      requireNamespace("dplyr", quietly = TRUE)) {
    d = st[1:250,] \%>\%
      dplyr::mutate(
        time = lubridate::as_datetime(time),
        time_end = dplyr::lead(time)
      ) \%>\%
      dplyr::filter(!is.na(time_end))
    raw = df \%>\% dplyr::filter(time >= min(d$time) & time <= max(d$time))
    dat = raw \%>\%
      tidyr::gather(axis, value, -time)
    d = d \%>\% dplyr::mutate(activity_y = as.numeric(sleep_wake == "sleep"))
    dat \%>\%
       ggplot2::ggplot(ggplot2::aes(x = time, y = value, colour = axis)) +
       ggplot2::geom_step() +
       ggplot2::geom_segment(
         data = d,
         ggplot2::aes(
           x = time, xend = time_end,
           y = activity_y, yend = activity_y,
           linetype = sleep_wake
         ),
         colour = "black", linewidth = 1, inherit.aes = FALSE
       ) +
       ggplot2::labs(linetype = "Sleep/wake")
  }

}
}
