% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asleep.R
\name{asleep}
\alias{asleep}
\title{Run \code{asleep} Model on Data}
\usage{
asleep(
  file,
  outdir = NULL,
  model_path = NULL,
  local_repo_path = NULL,
  min_wear_hours = 22L,
  time_shift = "0",
  report_light_and_temp = FALSE,
  pytorch_device = c("cpu", "cuda:0"),
  verbose = TRUE
)
}
\arguments{
\item{file}{accelerometry file to process, including CSV,
CWA, GT3X, and \code{GENEActiv} bin files}

\item{outdir}{output directory for CSVs and outputs}

\item{model_path}{the file path to the model.  If on disk, this can be
re-used and not re-downloaded.  If \code{NULL}, will download to the
temporary directory}

\item{local_repo_path}{path to load model definition from local repo description}

\item{min_wear_hours}{Min wear time in hours to be eligible for summary statistics computation.
The sleepnet paper uses 22}

\item{time_shift}{The number hours to shift forward or backward from
the current device time. e.g. +1 or -1}

\item{report_light_and_temp}{If true, it adds mean temp, and light columns to the predictions}

\item{pytorch_device}{device to use for prediction for PyTorch.}

\item{verbose}{print diagnostic messages}
}
\value{
A list of outputs, including summaries, paths, and dataframes.
}
\description{
Run \code{asleep} Model on Data
}
\examples{
file = system.file("extdata/example_sleep.csv.gz", package = "stepcount")
if (asleep_check()) {
  out = asleep(file = file)
  pred = out$predictions
}
\dontrun{
  file = system.file("extdata/example_sleep.csv.gz", package = "stepcount")
  df = readr::read_csv(file)
  if (asleep_check()) {
    out = stepcount(file = df)
    st = out$step_times
  }
  if (requireNamespace("ggplot2", quietly = TRUE) &&
      requireNamespace("tidyr", quietly = TRUE) &&
      requireNamespace("dplyr", quietly = TRUE)) {
    dat = df[10000:12000,] \%>\%
      dplyr::select(-annotation) \%>\%
      tidyr::gather(axis, value, -time)
    st = st \%>\%
      dplyr::mutate(time = lubridate::as_datetime(time)) \%>\%
      dplyr::as_tibble()
    st = st \%>\%
      dplyr::filter(time >= min(dat$time) & time <= max(dat$time))
    dat \%>\%
      ggplot2::ggplot(ggplot2::aes(x = time, y = value, colour = axis)) +
      ggplot2::geom_line() +
      ggplot2::geom_vline(data = st, ggplot2::aes(xintercept = time))
  }

}
}
