% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asleep.R
\name{have_asleep}
\alias{have_asleep}
\alias{asleep_check}
\alias{asleep_version}
\title{Check the \code{asleep} Python Module}
\usage{
have_asleep()

asleep_check()

asleep_version()
}
\value{
A logical value indicating whether the \code{asleep} Python module is available.
}
\description{
Check the \code{asleep} Python Module
}
\examples{
if (have_asleep()) {
   asleep_version()
}
}
